requirements.txt
asgiref==3.8.1
Django>=3.7.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
sqlparse==0.5.3